feat: use config log path and deal with logs before its initialization
This commit changes the new logging logic so logs are stored in the path
defined at patch-hub's config. Since logging should encompass the whole
application lifecycle, we have to correctly deal with logs produced
before/during Config initialization.
To solve this firstly we save the logs to a temporary path, and then,
after Config intitialization we move the existing logs to the path defined
at the Config and reload the logging layer so new logs are stored in
the updated path.

Signed-off-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>