refactor: move garbage_collector and log_on_error location
The garbage_collector function continues to work with the new
logging methods since it just depends on log path and file names
which didn't change. Temporary logs are on /tmp which is already
a temporary directory, so garbage_collector doesn't need to deal
with them.
The log_on_error macro was already adapted to deal with tracing::Level
log level.

Signed-off-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>