refactor: use mod.rs to clean up directory structure
Using `mod.rs` allows for a cleaner directory as we can avoid
having a file for each module in the root directory. This makes it
easier to identify the important files and represents the
compartmentalization of the code better.

Additionally, imports (of modules, external crates and std
create) in files now follow the recommended order.

Signed-off-by: Ivin Joel Abraham <ivinjabraham@gmail.com>