fix: init_details_actions checks if patch exists at expected path
Seeking to fix #69, this commit introduces a check to the function
init_details_actions. We hoped to stop the crash described in #69
by throwing an Err and catching it in the handle_latest_patchsets
fucntion before showing a popup to the user. For some still unkown
reason, the TUI freezes and 'melts' after the popup is shown.

Co-developed-by: Bruno Stephan <bruno.stephan@usp.br>
Signed-off-by: Bruno Stephan <bruno.stephan@usp.br>
Co-developed-by: Andre de Lima <aschwarz@usp.br>
Signed-off-by: Andre de Lima <aschwarz@usp.br>
Signed-off-by: Arthur Pilone <art.pilone@gmail.com>