fix: handles unexpected missing path with new enum
This commit fixes the previous one. The terminal
was unresponsive because the macro received an Err.

We created a new enum to test whether the init_details_action
function found a patch in the expected path. This enum is
used to test whether the file exists in the
handle_latest_patchsets function still using an Ok().

Fixes: #69

Co-developed-by: Bruno Stephan <bruno.stephan@usp.br>
Signed-off-by: Bruno Stephan <bruno.stephan@usp.br>
Co-developed-by: Andre de Lima <aschwarz@usp.br>
Signed-off-by: Andre de Lima <aschwarz@usp.br>
Signed-off-by: Arthur Pilone <art.pilone@gmail.com>