fix(release): disable CI job for release based on `cargo dist`
Once again, in trying to release the new version of `patch-hub`,
I stumbled upon an error in the CI, this time, with `cargo dist` [1]. As
mentioned in this issue [2], `cargo dist` deprecated support to the
Ubuntu 20.04 runner, which is the one used in the workflow. I have been
attempting to install other versions of `cargo dist`, but it seems that
there is an upstream compilation error. Besides, its own online book [3]
seems offline for the last days.

Due to this, disable CI job for releasing `patch-hub` using `cargo
dist`. This doesn't removes the job, just makes it not run at all. The
plan in-house a CI job for the release in the future.

[1]: https://github.com/kworkflow/patch-hub/actions/runs/15024335425
[2]: https://github.com/actions/runner-images/issues/11101
[3]: https://opensource.axo.dev/cargo-dist/book

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>