docs: add issue tracking subsection into CONTRIBUTING.md
This commit adds a new section suggesting that the users should mention the issue related
to the pull requests in the commentary, so github will automatically notify the issue that
there is a pull request on hold for that, avoiding double work around it.

As we are working on a free software environment, managing issues and contribution
is a hard task, most of the times, there may have some pull requests closing some
issue that haven't been revewed, and as it's not the most common approach search
in a long list of pull requests if an issue is still on hold, there may be cases when
other contributors may start to work on a problem that could be already solved.

Signed-off-by: JGBSouza <joaosouzaaa12@usp.br>
Reviewed-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>
Signed-off-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>