add: documentation: contribution: add more context on issue reporting
Tracking implementations as issues in github may be very usefull for a free software
context, as it can give more visibility of what is on development and let the others
be able to discuss an implementation even before it starts, making it less likely that
there will be an effort on developing a solution that will not be needed. In this sense
add a few more context on issue tracking section, stating the need to always open a issue
when there is no issue created for a future development.

Signed-off-by: JGBSouza <joaosouzaaa12@usp.br>