paper: add Hannah as one of the authors
Besides the JOSS's documentation stating that it is imperative to
include the all (no more, no less) authors that were involved in the
software being described, Hannah is the second person who most
contributed to the project. Adding her is a nobrainer and the correct
thing to do. In this respect, add Hannah to the author's list.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>